Ask the user to enable privileged mode if the docker version is unsupported
There's some kind of permission error when mounting the webDAV
filesystem in the application containers if the docker version
is greater than 1.10. we checked seccomp, capabilities, disabled
AppArmor and SELinux where applicable but the error still happens.

The applications run fine when using privileged mode, but that's
insecure and not recommended.